Moved zip file writer into zip storage source file #102
