+++ tests/storage/zip_file.py
@@ -6,12 +6,14 @@ import os
 import unittest
 import zipfile
 
+from plaso.engine import queue
 from plaso.formatters import manager as formatters_manager
 from plaso.formatters import mediator as formatters_mediator
 from plaso.lib import event
 from plaso.lib import eventdata
 from plaso.lib import timelib
 from plaso.formatters import winreg   # pylint: disable=unused-import
+from plaso.multi_processing import multi_process
 from plaso.serializer import protobuf_serializer
 from plaso.storage import time_range
 from plaso.storage import zip_file
@@ -292,7 +294,8 @@ class StorageFileTest(test_lib.StorageTestCase):
     test_event_tags = self._CreateTestEventTags()
 
     storage_file = zip_file.StorageFile(path)
-    storage_file.AddEventObjects(test_event_objects)
+    for test_event_object in test_event_objects:
+      storage_file.AddEventObject(test_event_object)
 
     storage_file.StoreTagging(test_event_tags[:-1])
     storage_file.StoreTagging(test_event_tags[-1:])
@@ -368,18 +371,6 @@ class StorageFileTest(test_lib.StorageTestCase):
 
       for test_event_object in test_event_objects:
         storage_file.AddEventObject(test_event_object)
-
-      storage_file.Close()
-
-  def testAddEventObjects(self):
-    """Tests the AddEventObjects function."""
-    test_event_objects = test_lib.CreateTestEventObjects()
-
-    with shared_test_lib.TempDirectory() as temp_directory:
-      temp_file = os.path.join(temp_directory, u'plaso.db')
-      storage_file = zip_file.StorageFile(temp_file)
-
-      storage_file.AddEventObjects(test_event_objects)
 
       storage_file.Close()
 
@@ -540,7 +531,8 @@ class StorageFileTest(test_lib.StorageTestCase):
       temp_file = os.path.join(temp_directory, u'plaso.db')
       storage_file = zip_file.StorageFile(temp_file)
 
-      storage_file.AddEventObjects(test_event_objects)
+      for test_event_object in test_event_objects:
+        storage_file.AddEventObject(test_event_object)
 
       storage_file.StoreTagging(test_event_tags[:-1])
       storage_file.StoreTagging(test_event_tags[-1:])
@@ -675,6 +667,47 @@ class ZIPStorageFileReaderTest(test_lib.StorageTestCase):
         1392438730000000]
 
     self.assertEqual(sorted(timestamps), expected_timestamps)
+
+
+class ZIPStorageFileWriterTest(unittest.TestCase):
+  """Tests for the ZIP-based storage file writer object."""
+
+  def testStorageWriter(self):
+    """Test the storage writer."""
+    test_event_objects = test_lib.CreateTestEventObjects()
+
+    # The storage writer is normally run in a separate thread.
+    # For the purpose of this test it has to be run in sequence,
+    # hence the call to WriteEventObjects after all the event objects
+    # have been queued up.
+
+    # TODO: add upper queue limit.
+    # A timeout is used to prevent the multi processing queue to close and
+    # stop blocking the current process.
+    test_queue = multi_process.MultiProcessingQueue(timeout=0.1)
+    test_queue_producer = queue.ItemQueueProducer(test_queue)
+    test_queue_producer.ProduceItems(test_event_objects)
+
+    test_queue_producer.SignalAbort()
+
+    preprocessing_object = event.PreprocessObject()
+
+    with shared_test_lib.TempDirectory() as temp_directory:
+      temp_file = os.path.join(temp_directory, u'plaso.db')
+      storage_writer = zip_file.ZIPStorageFileWriter(
+          test_queue, temp_file, preprocessing_object)
+      storage_writer.WriteEventObjects()
+
+      storage_file = zipfile.ZipFile(
+          temp_file, mode='r', compression=zipfile.ZIP_DEFLATED)
+
+      expected_filename_list = [
+          u'information.dump', u'plaso_index.000001', u'plaso_meta.000001',
+          u'plaso_proto.000001', u'plaso_timestamps.000001', u'serializer.txt']
+
+      filename_list = sorted(storage_file.namelist())
+      self.assertEqual(len(filename_list), 6)
+      self.assertEqual(filename_list, expected_filename_list)
 
 
 if __name__ == '__main__':
